pyproject.toml
[tool.poetry]
name = "gth-backend"
version = "0.1.0"
description = "GTH Scientia TI FastAPI Backend"
authors = ["Rhalf Augusto <rhalf@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"

fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }

sqlalchemy = "^2.0.30"
pydantic = "^2.7.0"

python-jose = "^3.3.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.9"
email-validator = "^2.1.0"
aiofiles = "^23.2.1"
lxml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
